+++ tests/spec/crm/objects/test_calls.py
@@ -1,5 +1,5 @@
 from hubspot import HubSpot
-from hubspot.crm.objects.calls import BasicApi, AssociationsApi, BatchApi, SearchApi
+from hubspot.crm.objects.calls import BasicApi, AssociationsApi, BatchApi, SearchApi, PublicObjectApi
 
 
 def test_is_discoverable():
@@ -8,3 +8,4 @@ def test_is_discoverable():
     assert isinstance(apis.associations_api, AssociationsApi)
     assert isinstance(apis.batch_api, BatchApi)
     assert isinstance(apis.search_api, SearchApi)
+    assert isinstance(apis.public_object_api, PublicObjectApi)