+++ tests/spec/crm/test_deals.py
@@ -1,5 +1,5 @@
 from hubspot import HubSpot
-from hubspot.crm.deals import BasicApi, AssociationsApi, BatchApi, SearchApi
+from hubspot.crm.deals import BasicApi, AssociationsApi, BatchApi, SearchApi, PublicObjectApi
 
 
 def test_is_discoverable():
@@ -8,4 +8,5 @@ def test_is_discoverable():
     assert isinstance(apis.basic_api, BasicApi)
     assert isinstance(apis.batch_api, BatchApi)
     assert isinstance(apis.search_api, SearchApi)
+    assert isinstance(apis.public_object_api, PublicObjectApi)
     assert hasattr(apis, "get_all")