+++ tests/spec/crm/test_objects.py
@@ -1,5 +1,5 @@
 from hubspot import HubSpot
-from hubspot.crm.objects import BasicApi, AssociationsApi, BatchApi, SearchApi, GDPRApi
+from hubspot.crm.objects import BasicApi, AssociationsApi, BatchApi, SearchApi, GDPRApi, PublicObjectApi
 
 
 def test_is_discoverable():
@@ -9,3 +9,4 @@ def test_is_discoverable():
     assert isinstance(apis.batch_api, BatchApi)
     assert isinstance(apis.search_api, SearchApi)
     assert isinstance(apis.gdpr_api, GDPRApi)
+    assert isinstance(apis.public_object_api, PublicObjectApi)