+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/HubSpot/hubspot-api-python/compare/v5.0.1...HEAD)
+## [Unreleased](https://github.com/HubSpot/hubspot-api-python/compare/v5.1.0...HEAD)
+
+## [5.1.0](https://github.com/HubSpot/hubspot-api-python/compare/v5.0.1...v5.1.0) - 2022-06-27
+
+### Added
+
+- CRM Public Object API clients
+
+### Fixed
+
+- returned id_property support for CRM API clients
 
 ## [5.0.1](https://github.com/HubSpot/hubspot-api-python/compare/v5.0.0...v5.0.1) - 2022-05-14